Move `TypeTranslation` into `CodeGenerator`
There are only a few clients of this code outside the code generator, and I plan to remove those in subsequent changed. This change moves the code around in preparation.
