+++ SpireAllSource.h
@@ -31,6 +31,5 @@
 #include "Source/SpireCore/SymbolTable.cpp"
 #include "Source/SpireCore/Syntax.cpp"
 #include "Source/SpireCore/TypeLayout.cpp"
-#include "Source/SpireCore/TypeTranslation.cpp"
 #include "Source/SpireCore/VariantIR.cpp"
 #include "Source/SpireLib/SpireLib.cpp"
